requirements.txt
Faker~=26.0.0
requests~=2.32.3
pytest~=8.3.1
allure-pytest